refactor: TaskBoardFileService 로직 수정
